add additional preview data
