+++ app/src/main/java/com/babylon/wallet/android/presentation/ui/composables/assets/TokensTab.kt
@@ -8,6 +8,7 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListScope
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.material3.HorizontalDivider
@@ -27,17 +28,28 @@ import com.babylon.wallet.android.presentation.ui.composables.ShimmeringView
 import com.babylon.wallet.android.presentation.ui.composables.Thumbnail
 import com.babylon.wallet.android.presentation.ui.modifier.throttleClickable
 import com.radixdlt.sargon.Decimal192
+import com.radixdlt.sargon.NetworkId
+import com.radixdlt.sargon.NonFungibleLocalId
+import com.radixdlt.sargon.PoolAddress
 import com.radixdlt.sargon.ResourceAddress
+import com.radixdlt.sargon.ValidatorAddress
 import com.radixdlt.sargon.annotation.UsesSampleValues
 import com.radixdlt.sargon.extensions.formatted
 import com.radixdlt.sargon.extensions.string
 import com.radixdlt.sargon.samples.sample
 import com.radixdlt.sargon.samples.sampleMainnet
+import com.radixdlt.sargon.samples.sampleRandom
+import rdx.works.core.domain.assets.AssetPrice
 import rdx.works.core.domain.assets.FiatPrice
+import rdx.works.core.domain.assets.NonFungibleCollection
+import rdx.works.core.domain.assets.PoolUnit
 import rdx.works.core.domain.assets.SupportedCurrency
 import rdx.works.core.domain.assets.Token
+import rdx.works.core.domain.assets.ValidatorWithStakes
 import rdx.works.core.domain.resources.ExplicitMetadataKey
+import rdx.works.core.domain.resources.Pool
 import rdx.works.core.domain.resources.Resource
+import rdx.works.core.domain.resources.Validator
 import rdx.works.core.domain.resources.metadata.Metadata
 import rdx.works.core.domain.resources.metadata.MetadataType
 
@@ -183,29 +195,146 @@ private fun TokenItem(
     }
 }
 
-@Preview
 @UsesSampleValues
-@Composable
-private fun TokenCardPreview() {
-    RadixWalletTheme {
-        AssetCard(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            TokenItem(
-                token = Token(
-                    Resource.FungibleResource(
-                        ResourceAddress.sampleMainnet(),
-                        Decimal192.sample(),
+internal val previewXrdToken = Token(
+    Resource.FungibleResource(
+        ResourceAddress.sampleMainnet(),
+        Decimal192.sample(),
+        metadata = listOf(
+            Metadata.Primitive(
+                ExplicitMetadataKey.NAME.key,
+                value = "XRD",
+                valueType = MetadataType.String
+            )
+        )
+    )
+)
+
+@Suppress("MagicNumber")
+@UsesSampleValues
+internal val previewAssetViewData = AssetsViewData(
+    xrd = previewXrdToken,
+    nonXrdTokens = listOf(
+        Token(
+            Resource.FungibleResource(
+                ResourceAddress.sampleRandom(NetworkId.MAINNET),
+                Decimal192.sample(),
+                metadata = listOf(
+                    Metadata.Primitive(
+                        ExplicitMetadataKey.NAME.key,
+                        value = "Token 1",
+                        valueType = MetadataType.String
+                    )
+                )
+            )
+        ),
+        Token(
+            Resource.FungibleResource(
+                ResourceAddress.sampleRandom(NetworkId.MAINNET),
+                Decimal192.sample(),
+                metadata = listOf(
+                    Metadata.Primitive(
+                        ExplicitMetadataKey.NAME.key,
+                        value = "Token 2",
+                        valueType = MetadataType.String
+                    )
+                )
+            )
+        )
+
+    ),
+    nonFungibleCollections = listOf(
+        NonFungibleCollection(
+            Resource.NonFungibleResource(
+                ResourceAddress.sampleMainnet(),
+                3L,
+                items = listOf(
+                    Resource.NonFungibleResource.Item(
+                        collectionAddress = ResourceAddress.sampleMainnet(),
+                        localId = NonFungibleLocalId.sample(),
                         metadata = listOf(
+                            Metadata.Primitive(ExplicitMetadataKey.NAME.key, "Item 1", MetadataType.String),
                             Metadata.Primitive(
-                                ExplicitMetadataKey.NAME.key,
-                                value = "XRD",
-                                valueType = MetadataType.String
+                                key = ExplicitMetadataKey.KEY_IMAGE_URL.key,
+                                "https://image-service-test-images.s3.eu-west-2.amazonaws.com/wallet_test_images/KL%20Haze-medium.jpg",
+                                valueType = MetadataType.Url
                             )
                         )
+                    ),
+                    Resource.NonFungibleResource.Item(
+                        collectionAddress = ResourceAddress.sampleMainnet(),
+                        localId = NonFungibleLocalId.sample(),
+                        metadata = listOf(Metadata.Primitive(ExplicitMetadataKey.NAME.key, "Item 2", MetadataType.String))
                     )
                 ),
-                fiatPrice = FiatPrice(Decimal192.sample(), SupportedCurrency.USD),
+                metadata = listOf(Metadata.Primitive(ExplicitMetadataKey.NAME.key, "Collection Name", MetadataType.String))
+            )
+        )
+    ),
+    epoch = null,
+    prices = mapOf(previewXrdToken to AssetPrice.TokenPrice(previewXrdToken, FiatPrice(Decimal192.sample(), SupportedCurrency.USD))),
+    poolUnits = listOf(
+        PoolUnit(
+            stake = Resource.FungibleResource(ResourceAddress.sampleMainnet(), ownedAmount = Decimal192.sample()),
+            pool = Pool(
+                address = PoolAddress.sampleMainnet(),
+                resources = listOf(
+                    Resource.FungibleResource(
+                        address = ResourceAddress.sampleRandom(NetworkId.MAINNET),
+                        ownedAmount = Decimal192.sample()
+                    )
+                ),
+                metadata = listOf(Metadata.Primitive(ExplicitMetadataKey.NAME.key, "Pool 1", MetadataType.String))
+            )
+        ),
+        PoolUnit(
+            stake = Resource.FungibleResource(ResourceAddress.sampleMainnet(), ownedAmount = Decimal192.sample()),
+            pool = Pool(
+                address = PoolAddress.sampleMainnet(),
+                resources = listOf(
+                    Resource.FungibleResource(
+                        address = ResourceAddress.sampleRandom(NetworkId.MAINNET),
+                        ownedAmount = Decimal192.sample()
+                    )
+                ),
+                metadata = listOf(Metadata.Primitive(ExplicitMetadataKey.NAME.key, "Pool 2", MetadataType.String))
+            )
+        )
+    ),
+    validatorsWithStakes = listOf(
+        ValidatorWithStakes(
+            validator = Validator(
+                address = ValidatorAddress.sampleMainnet(),
+                totalXrdStake = Decimal192.sample(),
+                stakeUnitResourceAddress = ResourceAddress.sampleMainnet(),
+                claimTokenResourceAddress = ResourceAddress.sampleMainnet(),
+                metadata = listOf(
+                    Metadata.Primitive(ExplicitMetadataKey.NAME.key, "Validator 1", MetadataType.String)
+                )
+            )
+        ),
+        ValidatorWithStakes(
+            validator = Validator(
+                ValidatorAddress.sampleRandom(NetworkId.MAINNET),
+                Decimal192.sample(),
+                ResourceAddress.sampleMainnet(),
+                ResourceAddress.sampleMainnet(),
+                metadata = listOf(
+                    Metadata.Primitive(ExplicitMetadataKey.NAME.key, "Validator 2", MetadataType.String)
+                )
+            )
+        )
+    )
+)
+
+@Preview(showBackground = true)
+@UsesSampleValues
+@Composable
+private fun TokensTabPreview() {
+    RadixWalletTheme {
+        LazyColumn {
+            tokensTab(
+                assetsViewData = previewAssetViewData,
                 isLoadingBalance = false,
                 action = AssetsViewAction.Click(
                     onFungibleClick = {},
